requirements.txt
appdirs==1.4.4
atomicwrites==1.4.0
attrs==20.2.0
black==20.8b1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.3
docopt==0.6.2
from==0.1.0
idna==2.10
import==0.0.1
importlib-metadata==1.7.0
iniconfig==1.0.1
more-itertools==8.5.0
mypy-extensions==0.4.3
packaging==20.4
path==15.0.0
pathspec==0.8.1
pluggy==0.13.1
psycopg2==2.8.6
py==1.9.0
pyodbc==4.0.30
pyparsing==2.4.7
pytest==6.0.2
python-decouple==3.3
python-slugify==4.0.1
regex==2020.11.13
requests==2.24.0
six==1.15.0
text-unidecode==1.3
toml==0.10.1
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.25.10
zipp==3.1.0
